requirements.txt

absl-py==0.11.0
aiohttp==3.7.3
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.6
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
discord.py==1.6.0
ffmpeg-python==0.2.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
isort==5.6.4
joblib==1.1.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.5.2.54
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-git==2018.2.1
pytz==2021.3
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==1.0.1
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.3
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.5.16